requirements.txt
mysql-connector-python-rf==2.2.2
numpy==1.24.2
